Upload Contains Duplicate.java
Let's see this code step by step :

Class Definition: The code defines a class named "Solution". This class contains a single method called containsDuplicate that takes an array of integers (int[] nums) as input and returns a boolean (true or false) indicating whether the array contains any duplicate numbers.

HashSet Declaration: Inside the method, a new HashSet named hm (short for "hash map") is created. A HashSet is a special type of collection that stores unique elements, meaning it doesn't allow duplicates. In this case, it will be used to store the numbers from the input array without duplicates.

Looping through the Array: The code uses a "for" loop to go through each number in the input array (nums). The loop variable i starts from 0 and goes up to one less than the length of the array.

Adding to the HashSet: Inside the loop, each number nums[i] is added to the HashSet hm. Since a HashSet only keeps unique values, if a number is already present in the HashSet, adding it again won't have any effect.

Checking for Duplicates: After adding all numbers to the HashSet, the code checks whether the size of the HashSet (hm.size()) is equal to the length of the input array (nums.length). If they are equal, it means that there are no duplicates in the array, because the HashSet only contains unique values. In this case, the method returns false, indicating that there are no duplicates.

Returning Result: If the sizes are not equal, it means that there are duplicates in the array, because the HashSet couldn't store all the values without any duplicates. In this case, the method returns true, indicating that there are duplicates.

So, in simpler terms, this code checks whether an array of numbers has any duplicate values. It does this by adding the numbers to a HashSet, which automatically removes duplicates. If the size of the HashSet is the same as the length of the original array, it means there are no duplicates. If the sizes are different, there are duplicates.